Move `BaseSpec` and relevant base types to `spec/base.ts`
